Add mv to rename branches.
Very similar to push. To rename protected remote branches, it'll create
a pull request; that pull request will handle naming the new branch,
but it will not delete the old one; in other words, for protected
branches, it's just as efficient to simply rename locally and then push
that branch (and delete the branch on the remote after unprotecting it).