Make sure to pass the branch name.
The branch is passed as None only if calling via CLI. If we call the
fetch function locally, we are not making use of Typer's default
parameters. In this case, we need to actually put a value for e.g. the
branch parameter, or actually set it to none, which would be Typer's
default also.